requirements.txt
requests==2.27.1
types-requests==2.27.25
bs4==0.0.1
flake8==3.9.2
tox==3.24.3
pytest==6.2.5
pytest-cov==2.12.1
mypy===0.910
click==8.1.3
iso639==0.1.4
python-dotenv==0.20.0